refactor: Photo service, update search function, and fix LazyInitializationException
Various refactorings are applied to PhotoService, breaking down its multiple responsibilities into separate services. The search function is also optimized to support searching either by tags or description in one function. The FetchType is changed to LAZY in PhotoEntity to fix the LazyInitializationException. The TagDto and PhotoDto conversion methods are improved for better mapping. The handling of TagEntity creation or updating is moved to TagService for better responsibility segregation. The explore page in the front-end is updated to reflect these backend changes.